refactor(goods): 상품 목록 페이지를 pages 레이어로 분리하고 page.tsx에 적용
