+++ src/app/(house)/page.tsx
@@ -1,10 +1,8 @@
 import { HydrationBoundary, dehydrate } from "@tanstack/react-query";
 import { GetGoodListRequestDto } from "@/entities/goods/goods.types";
 import { goodListInfiniteQueryOptions } from "@/features/goods/lib/goodListInfiniteQueryOptions";
-import { FilterSection } from "@/widgets/goods/FilterSection";
-import { GoodListSection } from "@/widgets/goods/GoodListSection";
+import { GoodListPage } from "@/pages/goods/ui/GoodListPage";
 import { getQueryClient } from "@/shared/lib/react-query";
-import * as s from "./style.css";
 
 type Props = {
   searchParams: Promise<Partial<Pick<GetGoodListRequestDto, "order">>>;
@@ -20,15 +18,8 @@ export default async function Home({ searchParams }: Props) {
   );
 
   return (
-    <main>
-      <div className={s.filterSection}>
-        <FilterSection />
-      </div>
-      <section className={s.mainGoodsSection}>
-        <HydrationBoundary state={dehydrate(queryClient)}>
-          <GoodListSection order={order} />
-        </HydrationBoundary>
-      </section>
-    </main>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <GoodListPage order={order} />
+    </HydrationBoundary>
   );
 }
