+++ src/widgets/goods/FilterSection/style.css.ts
@@ -1,9 +1,10 @@
 import { f } from "@/shared/styles/functions";
 import { responsiveStyle } from "@/shared/styles/functions/layout.css";
-import { CONTENT_WIDTH } from "@/shared/styles/globals.css";
+import {
+  CONTENT_WIDTH,
+  ORDER_FILTER_HEIGHT,
+} from "@/shared/styles/globals.css";
 import { style } from "@vanilla-extract/css";
-
-export const ORDER_FILTER_HEIGHT = 70;
 
 export const wrapper = style([
   f.flex,
